pyproject.toml
[project]
name = "autotrx"
version = "0.1.0"
description = "Personal Finance ETL with Human-in-the-Loop"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.1,<0.3.0)",
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
